UI, Rotation, and fixes
Adaptive UI now displays the name of the part the player is holding

Custom rotation script to override the rotation tracker of the XRGI.

Since it's no longer dependent on the controller's transform, it won't snap to the controller's rotation values when users grab them. Therefore the UnParent script is no longer needed.